feat(k8s/util) move ServiceTag for k8s Service to utils
